+++ src/commands/AuditController.php
@@ -7,6 +7,7 @@ use bedezign\yii2\audit\components\panels\Panel;
 use bedezign\yii2\audit\models\AuditEntry;
 use bedezign\yii2\audit\models\AuditError;
 use Yii;
+use yii\base\Exception;
 use yii\console\Controller;
 use yii\helpers\Console;
 use yii\helpers\Html;
@@ -26,6 +27,11 @@ class AuditController extends Controller
     public $entry;
 
     /**
+     * @var bool True to cleanup solo AuditEntry records (no trail/mail/error/javascript).
+     */
+    public $entrySolo;
+
+    /**
      * @var string|null Comma separated list of panels to cleanup.
      */
     public $panels;
@@ -42,7 +48,7 @@ class AuditController extends Controller
     {
         return array_merge(
             parent::options($actionID),
-            ($actionID == 'cleanup') ? ['entry', 'panels', 'age'] : []
+            ($actionID == 'cleanup') ? ['entry', 'entrySolo', 'panels', 'age'] : []
         );
     }
 
@@ -55,10 +61,14 @@ class AuditController extends Controller
     {
         /** @var Audit $audit */
         $audit = Yii::$app->getModule(Audit::findModuleIdentifier());
-        $panels = $this->panels !== null ? explode(',', $this->panels) : array_keys($audit->panels);
+        if ($this->panels === '') {
+            $panels = [];
+        } else {
+            $panels = !empty($this->panels) ? explode(',', $this->panels) : array_keys($audit->panels);
+        }
 
         // summary
-        $this->preCleanupSummary($this->entry, $panels, $this->age);
+        $this->preCleanupSummary($this->entry, $this->entrySolo, $panels, $this->age);
 
         // confirm
         if ($this->confirm('Cleanup the above data?')) {
@@ -76,6 +86,13 @@ class AuditController extends Controller
                     return self::EXIT_CODE_ERROR;
                 }
             }
+            // cleanup solo audit_entry
+            if ($this->entrySolo) {
+                if (!$this->cleanupEntrySolo()) {
+                    $this->stdout("\nCleanup failed.\n", Console::FG_RED);
+                    return self::EXIT_CODE_ERROR;
+                }
+            }
             // success!
             $this->stdout("\nCleanup was successful.\n", Console::FG_GREEN);
         }
@@ -86,10 +103,11 @@ class AuditController extends Controller
      * Displays a summary of the data and dates to clean
      *
      * @param bool $entry
+     * @param bool $entrySolo
      * @param array $panels
      * @param int|null $maxAge
      */
-    protected function preCleanupSummary($entry, $panels, $maxAge)
+    protected function preCleanupSummary($entry, $entrySolo, $panels, $maxAge)
     {
         $audit = Audit::getInstance();
 
@@ -120,6 +138,11 @@ class AuditController extends Controller
             $this->stdout("\t" . 'AuditEntry .............. ' . $date . "\n");
         }
 
+        // audit entry solo
+        if ($entrySolo) {
+            $this->stdout("\t" . 'AuditEntry solo ......... ' . date('Y-m-d 23:59:59') . "\n");
+        }
+
         $this->stdout("\n");
     }
 
@@ -148,6 +171,41 @@ class AuditController extends Controller
         $time = microtime(true) - $start;
         $this->stdout("\n*** failed to clean AuditEntry (time: " . sprintf("%.3f", $time) . "s)\n", Console::FG_RED);
         return false;
+    }
+
+    /**
+     * Cleans the AuditEntry solo data (no trail/mail/error/javascript)
+     *
+     * @return bool
+     */
+    protected function cleanupEntrySolo()
+    {
+        $this->stdout("\n*** cleaning AuditEntry solo", Console::FG_YELLOW);
+        $start = microtime(true);
+        $count = 0;
+        foreach (AuditEntry::find()->each(100) as $auditEntry) {
+            /** @var AuditEntry $auditEntry */
+            /** @var Audit $audit */
+            $audit = Yii::$app->getModule('audit');
+            $auditEntryCurrent = $audit->getEntry();
+            if ($auditEntryCurrent && $auditEntryCurrent->id == $auditEntry->id) {
+                continue;
+            }
+            if (!$auditEntry->errors && !$auditEntry->javascripts && !$auditEntry->mails && !$auditEntry->trails) {
+                foreach ($auditEntry->data as $data) {
+                    $data->delete();
+                }
+                try {
+                    $auditEntry->delete();
+                    $count++;
+                    $this->stdout('.', Console::FG_CYAN);
+                } catch (Exception $e) {
+                }
+            }
+        }
+        $time = microtime(true) - $start;
+        $this->stdout("\n*** cleaned AuditEntry (records: " . $count . ",time: " . sprintf("%.3f", $time) . "s)\n", Console::FG_GREEN);
+        return true;
     }
 
     /**